src/api/v1/django-project/backend/web-back/requirements.txt
asgiref==3.4.0
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.2
gunicorn==20.1.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
python-dotenv==0.18.0
pytz==2021.1
sqlparse==0.4.1
